created landscape utils to remove duplicated code - everything that might be used in other landscape components are now utils functions - removed style-helper util and moved both functions fromt it into landscape utils
