+++ src/main/app/src/Components/LandscapeComponent/LandscapeItem/LandscapeItem.scss
@@ -60,6 +60,7 @@ $dotSize: 50px;
     padding-bottom: 16px;
     border-bottom: 1px solid $border-color;
 
+    .labelContent,
     .item {
       margin-top: 16px;
     }
